+++ backend/mora/mapping.py
@@ -34,6 +34,7 @@ EMPLOYEE = "employee"
 LOCATION = 'location'
 ERROR = 'error'
 USER_SETTINGS = 'user_settings'
+INTEGRATION_DATA = 'integration_data'
 
 # Address
 ADDRESS_KEY = 'Adresse'