+++ backend/mora/service/__init__.py
@@ -21,6 +21,7 @@ from . import manager  # noqa
 from . import org
 from . import orgunit
 from . import role  # noqa
+from . import integration_data
 
 blueprints = (
     address.blueprint,
@@ -33,4 +34,5 @@ blueprints = (
     org.blueprint,
     orgunit.blueprint,
     exports.blueprint,
+    integration_data.blueprint,
 )