+++ frontend/src/modules/timeMachine/index.vue
@@ -4,11 +4,11 @@
 
     <div class="row">
       <div class="col">
-        <mo-time-machine-column/>
+        <mo-time-machine-column store-id="$_timeMachineLeft"/>
       </div>
 
       <div class="col">
-        <mo-time-machine-column/>
+        <mo-time-machine-column store-id="$_timeMachineRight"/>
       </div>
     </div>
   </div>